requirements.txt
blis==0.7.4
catalogue==2.0.4
certifi==2021.5.30
charset-normalizer==2.0.3
click==7.1.2
colorama==0.4.4
cymem==2.0.5
idna==3.2
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
murmurhash==1.0.5
mysql-connector==2.2.9
mysql-connector-python==8.0.26
nltk==3.6.2
numpy==1.21.1
packaging==21.0
pathy==0.6.0
preshed==3.0.5
protobuf==3.17.3
pydantic==1.8.2
pyparsing==2.4.7
regex==2021.7.6
requests==2.26.0
six==1.16.0
smart-open==5.1.0
spacy==3.1.1
spacy-legacy==3.0.8
srsly==2.4.1
thinc==8.0.8
tqdm==4.61.2
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.6
wasabi==0.8.2
